Renamed UIFactory to GameObjectFactory, fixed the issue with lobby retrieval exceptions (parsing error)

+++ Assets/BossRoom/Scripts/Shared/Net/UnityServices/Lobbies/LobbyContentHeartbeat.cs
@@ -173,7 +173,7 @@ namespace BossRoom.Scripts.Shared.Net.UnityServices.Lobbies
             data.Add("State", ((int)lobby.State).ToString()); // Using an int is smaller than using the enum state's name.
             data.Add("State_LastEdit", lobby.Data.State_LastEdit.ToString());
             data.Add("RelayNGOCode_LastEdit", lobby.Data.RelayNGOCode_LastEdit.ToString());
-            data.Add("OnlineMode", lobby.Data.OnlineMode.ToString());
+            data.Add("OnlineMode", ((int)lobby.Data.OnlineMode).ToString());
 
             return data;
         }
